Silence warnings

+++ menu/cbs/menu_cbs_right.c
@@ -385,9 +385,9 @@ static int action_right_video_gpu_index(unsigned type, const char *label,
 
          if (list)
          {
-            if (settings->ints.vulkan_gpu_index < list->size - 1)
+            if (settings->ints.vulkan_gpu_index <(int)(list->size - 1))
                settings->ints.vulkan_gpu_index++;
-            else if (settings->ints.vulkan_gpu_index == list->size - 1)
+            else if (settings->ints.vulkan_gpu_index == (int)(list->size - 1))
                settings->ints.vulkan_gpu_index = 0;
          }
 
@@ -401,9 +401,9 @@ static int action_right_video_gpu_index(unsigned type, const char *label,
 
          if (list)
          {
-            if (settings->ints.d3d10_gpu_index < list->size - 1)
+            if (settings->ints.d3d10_gpu_index < (int)(list->size - 1))
                settings->ints.d3d10_gpu_index++;
-            else if (settings->ints.d3d10_gpu_index == list->size - 1)
+            else if (settings->ints.d3d10_gpu_index == (int)(list->size - 1))
                settings->ints.d3d10_gpu_index = 0;
          }
